fwffs

+++ backend/src/main/java/com/Proyecto_Datos/Proyectodat/controller/ProduccionController.java
@@ -25,7 +25,7 @@ public class ProduccionController {
     }
 
     @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable Long id) {
+    public void eliminar(@PathVariable Integer id) {
         produccionService.eliminar(id);
     }
 }
