+++ backend/src/main/java/com/Proyecto_Datos/Proyectodat/controller/TurnoController.java
@@ -25,7 +25,7 @@ public class TurnoController {
     }
 
     @DeleteMapping("/{id}")
-    public void eliminar(@PathVariable Long id) {
+    public void eliminar(@PathVariable Integer id) {
         turnoService.eliminar(id);
     }
 }
